move implementation to paddle.random
