+++ python/paddle/random.py
@@ -14,6 +14,7 @@
 
 from __future__ import annotations
 
+import paddle
 from paddle.base import core
 
 __all__ = ["initial_seed"]
@@ -35,13 +36,56 @@ def initial_seed() -> int:
     return core.default_cpu_generator().initial_seed()
 
 
-def __getattr__(name):
-    if name == "get_rng_state":
-        import sys
-        from importlib import import_module
+def get_rng_state(
+    device: str | None = None,
+) -> list[core.GeneratorState]:
+    """
+    Get all random states of random generators of specified device.
 
-        mod = import_module("paddle", package=__name__)
-        attr = getattr(mod, "get_rng_state", mod)
-        setattr(sys.modules[__name__], name, attr)
-        return attr
-    raise AttributeError(f"module 'paddle.random' has no attribute '{name}'")
+    Args:
+        device(str): This parameter determines the specific running device.
+            It can be ``cpu``, ``gpu``, ``xpu``, Default is None.
+            If None, return the generators of current device (specified by ``set_device``).
+
+    Returns:
+        list[GeneratorState], object.
+
+    Examples:
+        .. code-block:: pycon
+
+            >>> import paddle
+            >>> sts = paddle.get_rng_state()
+    """
+    state_list = []
+    if device is None:
+        place = paddle.framework._current_expected_place_()
+    else:
+        place = paddle.device._convert_to_place(device)
+
+    if isinstance(place, paddle.CPUPlace):
+        state_list.append(core.default_cpu_generator().get_state())
+    elif isinstance(place, paddle.CUDAPlace):
+        for i in range(core.get_cuda_device_count()):
+            state_list.append(core.default_cuda_generator(i).get_state())
+    elif isinstance(place, paddle.XPUPlace):
+        for i in range(core.get_xpu_device_count()):
+            state_list.append(core.default_xpu_generator(i).get_state())
+    elif isinstance(place, paddle.CustomPlace):
+        dev_cnt = sum(
+            [
+                place.get_device_type() == s.split(':')[0]
+                for s in core.get_available_custom_device()
+            ]
+        )
+        for i in range(dev_cnt):
+            state_list.append(
+                core.default_custom_device_generator(
+                    core.CustomPlace(place.get_device_type(), i)
+                ).get_state()
+            )
+    else:
+        raise ValueError(
+            f"get_rng_state is not implemented for current device: {place}"
+        )
+
+    return state_list
